Исправляет всякий треш
