+++ js/form-states.js
@@ -1,5 +1,3 @@
-import {adForm} from './form-validation.js';
-
 const formElements = document.querySelectorAll('form');
 
 const setFormElementsDisableState = (form) => {
@@ -27,6 +25,3 @@ export const activateForms = (form) => {
   setFormElementsEnableState(form);
   form.classList.remove('disabled');
 };
-
-deactivateForms();
-activateForms(adForm);
